+++ Workspace/EWMS_project/EWMS/src/main/java/it/unisa/ewms/persistance/interfaces/ITaskDAO.java
@@ -1,18 +1,17 @@
 package it.unisa.ewms.persistance.interfaces;
 
-import it.unisa.ewms.persistance.beans.Task;
-import it.unisa.ewms.persistance.beans.Tipi;
-import it.unisa.ewms.persistance.beans.Utente;
+import it.unisa.ewms.model.beans.Task;
+import it.unisa.ewms.model.beans.Tipi;
+import it.unisa.ewms.model.beans.Utente;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public interface ITaskDAO {
     void create(Task task) throws Exception;
-    Task findById(int id) throws Exception;
+    Task findById(long id) throws Exception;
     List<Task> findByUtente(Utente utente) throws Exception;
     void update(Task task) throws Exception;
     void updateStatus(long id, Tipi.stato nuovoStato) throws Exception;
-    void delete(int id) throws Exception;
+    void delete(long id) throws Exception;
 
 }
